+++ packages/plugins/max-depth/src/index.ts
@@ -11,7 +11,7 @@ import {
 } from 'graphql';
 
 type MaxDepthOptions = { n?: number; ignoreIntrospection?: boolean };
-const maxDepthDefaultOptions: MaxDepthOptions = {
+const maxDepthDefaultOptions: Required<MaxDepthOptions> = {
   n: 6,
   ignoreIntrospection: true,
 };
@@ -20,12 +20,12 @@ class MaxDepthVisitor {
   public readonly OperationDefinition: Record<string, any>;
 
   private readonly context: ValidationContext;
-  private readonly options: MaxDepthOptions;
+  private readonly config: Required<MaxDepthOptions>;
   private onError: (msg: string) => any;
 
   constructor(context: ValidationContext, onError: (msg: string) => any, options?: MaxDepthOptions) {
     this.context = context;
-    this.options = Object.assign(
+    this.config = Object.assign(
       {},
       maxDepthDefaultOptions,
       ...Object.entries(options ?? {}).map(([k, v]) => (v === undefined ? {} : { [k]: v })),
@@ -39,28 +39,29 @@ class MaxDepthVisitor {
 
   onOperationDefinitionEnter(operation: OperationDefinitionNode): void {
     const depth = this.countDepth(operation);
-    if (depth > this.options.n!) {
+    if (depth > this.config.n) {
       this.onError('Query is too deep.');
     }
   }
 
   private countDepth(
     node: FieldNode | FragmentDefinitionNode | InlineFragmentNode | OperationDefinitionNode | FragmentSpreadNode,
-    depth: number = 0,
+    parentDepth: number = 0,
   ): number {
-    if (this.options.ignoreIntrospection && 'name' in node && node.name?.value === '__schema') {
+    if (this.config.ignoreIntrospection && 'name' in node && node.name?.value === '__schema') {
       return 0;
     }
+    let depth = parentDepth;
 
     if ('selectionSet' in node && node.selectionSet) {
       for (let child of node.selectionSet.selections) {
-        depth = Math.max(depth, this.countDepth(child, depth + 1));
+        depth = Math.max(depth, this.countDepth(child, parentDepth + 1));
       }
     }
     if (node.kind == Kind.FRAGMENT_SPREAD) {
       const fragment = this.context.getFragment(node.name.value);
       if (fragment) {
-        depth = Math.max(depth, this.countDepth(fragment, depth + 1));
+        depth = Math.max(depth, this.countDepth(fragment, parentDepth + 1));
       }
     }
     return depth;