+++ packages/plugins/block-field-suggestions/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @escape.tech/graphql-armor-block-field-suggestions
 
+## 1.2.3
+
+### Patch Changes
+
+- 0c5d3db: fix: readme package name
+
+## 1.2.2
+
+### Patch Changes
+
+- f8a2a9b: refactor: usage of Required type
+
+## 1.2.1
+
+### Patch Changes
+
+- 8e2f90c: feat(field-suggestion): mask usage
+
 ## 1.2.0
 
 ### Minor Changes